+++ ObsStrawket.Test/Utilities/TestUtil.cs
@@ -5,15 +5,6 @@ namespace ObsStrawket.Test.Utilities {
   public static class TestUtil {
     public const double Epsilon = 0.00000003;
 
-    public static object ToDoubleOrInt(object boxed) {
-      return boxed switch {
-        float => (double)(float)boxed,
-        byte => (int)(byte)boxed,
-        ushort => (int)(ushort)boxed,
-        _ => boxed,
-      };
-    }
-
     public static Task WhenAnyThrowsAsync(params Task[] tasks) {
       return WhenAnyThrowsAsync(tasks as IEnumerable<Task>);
     }